refactor(#196): dedup shared helpers surfaced by /simplify
Quality cleanup of the operation-redundancy work, no behavior change:

- Move `ATTRIBUTE_NAMESPACE` and the attribute/docblock shape detection to
  `AuthoredSchemaShape::detect()`, replacing the duplicated `shapeFor()` in both
  the schema and operation rules and inverting the `OaAttributeRemover` →
  rule-class dependency.
- Extract the duplicated `fileFor()` reflect-and-catch into a shared
  `ResolvesDeclaringFile` trait used by both fixers.
- Single shared route-key definition `LintContext::operationKey()`, called by
  both the runner's index builder and the rule, so the "{method} {uri}"
  convention can't drift.
- Correct the rule docblock / CHANGELOG: the harvester merges only metadata +
  responses; `parameters`/`requestBody` are checked as a conservative guard.
